Fixes on NavBar and Image Carousel set in appropriate place
