+++ src/components/Home.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import Navbar from "./Navbar";
+import ImageCarousel from "./ImageCarousel"; // Import the ImageCarousel component
+const Home = () => {
+    return (
+        <>
+            <Navbar />
+            <ImageCarousel />
+        </>
+    );
+};
+
+export default Home;
