+++ src/components/Plant.jsx
@@ -6,21 +6,21 @@ const plantsData = [
     id: 1,
     title: "AYURVEDA",
     description: "Ayurveda is a traditional Indian medical system that aims to restore balance to the body, mind, and spirit.",
-    image: "src/ayurveda image.jpg",
+    image: "src/assets/images/ayurveda image.jpg",
     system: "Ayurvedic Herbs",
   },
   {
     id: 2,
     title: "UNANI",
     description: "Unani medicine, also known as Yunani medicine, is a traditional system of healing and health maintenance that originated in ancient Greece and later developed in the Middle East and South Asia.",
-    image: "src/unani image.jpeg",
+    image: "src/assets/images/unani image.jpeg",
     system: "Unani Herbs",
   },
   {
     id: 3,
     title: "HOMOEOPATHY",
     description: "Homeopathy is a system of alternative medicine that is based on the principle of meaning that a substance that causes symptoms in a healthy person can, in very small doses, treat similar symptoms in a sick person.",
-    image: "src/Homeopathy image.jpeg",
+    image: "src/assets/images/Homeopathy image.jpeg",
     system: "Homeopathic Herbs",
   },
 ];
